requirements.txt
click==7.1.1
future==0.18.2
Jinja2==2.11.2
joblib==0.14.1
livereload==2.6.1
lunr==0.5.6
Markdown==3.2.1
MarkupSafe==1.1.1
mkdocs==1.1
nltk==3.5
PyYAML==5.3.1
regex==2020.4.4
six==1.14.0
tornado==6.0.4
tqdm==4.45.0
